pyproject.toml
[project]
name = "niji-pallet-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "pydantic-settings>=2.14.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "aiosqlite>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
    "ty>=0.0.32",
]

# ================================================================
# ty 設定
# ================================================================
[tool.ty.src]
include = ["src/core"]
exclude = ["src/generated", "tests"]

# ================================================================
# ruff 設定
# ================================================================
[tool.ruff]
line-length = 88
target-version = "py314"
exclude = ["src/generated"]

[tool.ruff.lint]
# E: pycodestyle / F: Pyflakes / I: isort / B: flake8-bugbear / UP: pyupgrade / N: pep8-naming
select = ["E", "F", "I", "B", "UP", "N"]
ignore = [
    "E501",  # 行長は formatter に委ねる
    "B008",  # デフォルト引数での関数呼び出し（FastAPI の Depends パターンで多用）
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # テストでの assert を許可

# ================================================================
# pytest 設定
# ================================================================
[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ================================================================
# pytest 設定
# ================================================================
[tool.datamodel-codegen]
input = "docs/openapi.yaml"
input-file-type = "openapi"
output = "src/generated/models/openapi_models.py"
output-model-type = "pydantic_v2.BaseModel"
target-python-version = "3.14"
use-standard-collections = true
use-union-operator = true
field-constraints = true
use-schema-description = true
collapse-root-models = true
formatters = ["ruff-format", "ruff-check"]